Utilise le serveur Web interne à PHP à la place de lighttpd sur les machines non Windows.

+++ sources/BellePoule/supervisor/contest.cpp
@@ -980,18 +980,16 @@ Contest::~Contest ()
   _state = State::LEAVING;
 
   // www
-#ifndef DEBUG
-  if (_filename)
+  if (_filename && Global::_www)
   {
     gchar *base_name = g_path_get_basename (_filename);
-    gchar *www_file  = g_build_filename (Global::_share_dir, "webserver", "LightTPD", "www", "cotcot", base_name, NULL);
+    gchar *www_file  = g_build_filename (Global::_www, base_name, NULL);
 
     g_unlink (www_file);
 
     g_free (base_name);
     g_free (www_file);
   }
-#endif
 
   g_free (_authority);
   g_free (_fie_id);
@@ -1837,16 +1835,16 @@ void Contest::Save ()
     Save (_filename);
 
     // www
-#ifndef DEBUG
+    if (Global::_www)
     {
       gchar *base_name = g_path_get_basename (_filename);
-      gchar *www_file  = g_build_filename (Global::_share_dir, "webserver", "LightTPD", "www", "cotcot", base_name, NULL);
+      gchar *www_file  = g_build_filename (Global::_www, base_name, NULL);
 
       Save (www_file);
+
       g_free (base_name);
       g_free (www_file);
     }
-#endif
 
     // Backup
     if (_tournament)
